Make it clearer that files can be dropped
A fullsize drop target is now shown over the page when a file is dragged
over it, to make it clearer that a file can be dropped there.

It also indicates better that the tool is handling the file and dropping it
on the text editor won't simply insert the content there (like it
usually works for textareas).