Allow and add more parsers
This commit slightly changes the internal structure so that it is easier
to add new parsers to the explorer. See the README for how to.

The esprima/babel buttons is now a dropdown and shows the currently
selected parser. The parser info in the top right corner now also links
to the homepage of the parser.

New parsers:

- acorn
- espree

babel was renamed to babylon