Add auto-completion to jscodeshift transformer (closes #92)
This uses tern to provide auto-completion for the jscodeshift, recast
and ast-types API.

This adds a new script (`scripts/definition-generator.js`) which
programmatically generates most of the API definitions for tern.

The integration is kind of hacky. I cloned the editor so that all the
tern dependencies are loaded only on demand. Unfortunately some of the
dependecies have to be set globally, otherwise tern will load its own
acorn copy which doesn't work properly with webpack.

There still needs to more descriptions added to the API methods. Hoping
for contributors :)