Fix uglify2 conflicts
Unfortunately it seems that npm3 is deduping our version of uglify-js,
which is loaded from GitHub, and webpack's version of uglify-js (loaded
from npm).
Both packages have a compatible version number, so it's not too
surprising. However, our version wins, which means webpack is also using
the harmony version, which results in issues because that version of the
parser is a bit broken. (the issue is that files are not minified in the
final build)

I'm solving this in the meantime by adding the uglify harmony branch as
submodule and referencing it directly in the parser file.