Switch to redux (#111)
* Switch to redux

First attempt to switch to redux and redux-sagas. I'm still using PubSub
to circumvent the whole system in some cases.

This can certainly be improved. I feel like the reducer is not doing
much more than setting values. Maybe more logic should be moved to the
reducer.

* Cleanup eslint

There seems to be a bug in npm or our setup. An `npm install
--no-shrinkwrap` doesn't install the version of `js-yaml` that eslint v1
needs (which is a problem for `npm shrinkwrap`). However, running `npm
install` *again* afterwards installs the missing dependency.