Properly hide type keys in different parsers
Initially the checkbox was implemented only with a hardcoded `type` key in mind despite various parsers using different properties for the `getNodeName`, which resulted in a checkbox not having any effect for these parsers.

This adds a new property `typeKeys: Set<string>` that can be used on parsers with custom `getNodeName` to override set of keys that should be hidden when the checkbox is enabled.