[dev] Upgrade eslint dev dependency
Yarn seems to have problems installing eslint4 and eslint since both
refer to the same package. The result is that there is no
`node_modules/eslint` folder and plugins fail.

Of course just upgrading to the latest version is not a long term
solution since we would run into the same issue if we add a `eslint6`
package. The real solution is decoupling parser/transformer bundles from
the website.