Split parser related code into separate bundle
The parsers themselves are already lazy loaded, but the parser adapters
are currently bundled with the app. The app doesn't change as often as
parser adapters do, so it makes sense to split them out.

Of course that still means that the whole file changes if a single
parser is updated, but that situation cannot really be improved until
parsers are build and bundled in isolation.