Transform code in middleware instead of component
The goal is to store version information of the parser/transform on the
parse/transform result instead of "hardcoding" it on the parser adapter.
This will allow us to decouple the actual parser package from the
website and adapter code.

This is likely not the final and perfectly solution for working with
parsers in astexplorer and it may not even be part of the final
solution, but it's a step towards it.