Improve autofocus
Scrolling to the right element didn't always work as expected because
instead of taking the center of the *visible* area, we took the center
of the whole element, which in this case was larger than the visible
part. This was fixed by adding overflow to the list element itself.

Furthermore, some parsers include additional (and often recursive)
information in the location object. When computing the focus, location
nodes should be ignored.