feat: Add support for Builder visualization
This adds support for a Builder tab in the visualization menu that will
generate code to recreate the original ast in the code editor using
calls to ast node constructors in the @babel/types library.

This can be very useful when writing codemods to quickly determine the
proper usage of @babel/types library to generate ast nodes.